Refactor: Move getFlippedPositions function to MakeMove.kt file
